update client reconciler
