+++ pkg/common/cluster_actions.go
@@ -12,6 +12,11 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
+
+  "encoding/json"
+  "net/http"
+  "io/ioutil"
+  "os"
 )
 
 var log = logf.Log.WithName("action_runner")
@@ -144,6 +149,36 @@ func (i *ClusterActionRunner) UpdateRealm(obj *v1alpha1.KeycloakRealm) error {
 	}
 	return i.keycloakClient.UpdateRealm(obj)
 }
+type Domain struct {
+	Code string
+	Id   string
+	Url  string
+}
+func retrieveDomains(obj *v1alpha1.KeycloakClient) []string {
+  response, err := http.Get(obj.Spec.ApiDomain)
+  if err != nil {
+    fmt.Print(err.Error())
+  }
+  responseData, err := ioutil.ReadAll(response.Body)
+  if err != nil {
+    log.Error(err, "")
+    os.Exit(1)
+  }
+  jsonString := string(responseData)
+  fmt.Println(jsonString)
+  jsonArray := []byte(responseData)
+  var domains []Domain
+  err2 := json.Unmarshal(jsonArray, &domains)
+  if err2 != nil {
+    fmt.Println("error:", err2)
+  }
+  retrievedDomains := []string{}
+  for k := range domains {
+  retrievedDomains = append(retrievedDomains, domains[k].Url)
+  }
+  return retrievedDomains
+}
+
 
 func (i *ClusterActionRunner) CreateClient(obj *v1alpha1.KeycloakClient, realm string) error {
 	if i.keycloakClient == nil {
@@ -157,6 +192,9 @@ func (i *ClusterActionRunner) CreateClient(obj *v1alpha1.KeycloakClient, realm s
 	}
 
 	obj.Spec.Client.ID = uid
+  if(obj.Spec.Client.ClientID == "subscriber-web" && len(obj.Spec.ApiDomain) != 0) {
+    obj.Spec.Client.RedirectUris = retrieveDomains(obj)
+  }
 
 	return i.client.Update(i.context, obj)
 }
@@ -164,6 +202,9 @@ func (i *ClusterActionRunner) CreateClient(obj *v1alpha1.KeycloakClient, realm s
 func (i *ClusterActionRunner) UpdateClient(obj *v1alpha1.KeycloakClient, realm string) error {
 	if i.keycloakClient == nil {
 		return errors.Errorf("cannot perform client update when client is nil")
+	}
+	if(obj.Spec.Client.ClientID == "subscriber-web" && len(obj.Spec.ApiDomain) != 0) {
+	  obj.Spec.Client.RedirectUris = retrieveDomains(obj)
 	}
 	return i.keycloakClient.UpdateClient(obj.Spec.Client, realm)
 }
